// src/interfaces/Shop.Interface.tsx
import {Subcategory} from './Subcategory.interface';

export interface CarItemProps {
  subcategory: Subcategory;
  cantidad: number;
}
// Generated by https://quicktype.io

export interface MyShopResponse {
  _id: string;
  car: CarItemProps[];
  user: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}

export interface Category {
  name: string;
  id: string;
}

export interface CantPaq {
  oneandhalfkgPrice: number;
  twokgPrice: number;
  threekgPrice: number;
  fourkgPrice: number;
  fivekgPrice: number;
  sixkgPrice: number;
  sevenkgPrice: number;
  eigthkgPrice: number;
  ninekgPrice: number;
  tenkgPrice: number;
}